LSR-851 Rename Interactor components
