+++ android/library/src/main/java/com/liferay/mobile/screens/gallery/interactor/BaseGalleryInteractor.java
@@ -1,0 +1,9 @@
+package com.liferay.mobile.screens.gallery.interactor;
+
+import com.liferay.mobile.screens.base.interactor.Interactor;
+
+/**
+ * @author Víctor Galán Grande
+ */
+public interface BaseGalleryInteractor extends Interactor<GalleryInteractorListener>{
+}
